Extract Evohome base entities to separate module
